api/pyproject.toml
[tool.ruff]
line-length = 88

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.poetry]
name = "virgilbookstore"
version = "0.1.0"
description = ""
authors = ["Pedro Fonseca"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
black = "^23.12.1"
pre-commit = "^3.6.0"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
ruff = "^0.1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
